backend/requirements.txt
# ============================================================
# Flux Backend — Python Dependencies
# ============================================================

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Data validation
pydantic==2.10.4
pydantic-settings==2.7.1

# Database (Supabase)
supabase==2.11.0

# AI / LLM
openai==1.59.3

# Utilities
python-dotenv==1.0.1
httpx==0.28.1

# Dev / Testing
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.8.6
